fix crash issue with newer dpdk

+++ app/cmdline.c
@@ -792,7 +792,7 @@ cmd_obj_lpm_lkp_parsed(void* parsed_result,
 {
 	struct cmd_obj_lpm_lkp_result *res = parsed_result;
 	uint32_t next_hop;
-	uint16_t next_hop6;
+	uint8_t next_hop6;
 	int i;
 	char buf[INET6_ADDRSTRLEN];
